Cleanup World Subcommands
