+++ bin/app/file_manager.d.ts
@@ -51,3 +51,16 @@ export declare function writeBufferFileFromString(path: string, data: string, ov
  * @param path the path to delete the file from
  */
 export declare function deleteFile(path: string): void;
+/**
+ * @remarks copies directory from source to destination recursively
+ * @param src the source directory
+ * @param dest the target directory
+ */
+export declare function copyDir(src: string, dest: string): void;
+/**
+ * @remarks compresses a directory to a zip-like file
+ * @param dir the directory to compress
+ * @param zipPath the path the compressed directory should be written to
+ * @param callback the callback to run when the compression finishes
+ */
+export declare function archiveDirToZip(dir: string, zipPath: string, callback: Function): void;
