+++ ts/app/file_manager.ts
@@ -3,6 +3,8 @@ import * as glob from 'glob';
 import * as Global from './globals';
 import * as Chalk from 'chalk';
 import * as JSONC from 'comment-json';
+import * as archiver from 'archiver';
+import path from 'path';
 
 const chalk = new Chalk.Instance();
 
@@ -163,4 +165,42 @@ function makeDirectory(path: string) {
         fs.mkdirSync(dir, {recursive: true});
         console.log(`${chalk.yellow(`Creating Directory at: ${dir}`)}`);
     }
+}
+
+/**
+ * @remarks copies directory from source to destination recursively
+ * @param src the source directory
+ * @param dest the target directory
+ */
+export function copyDir(src: string, dest: string)
+{
+    fs.mkdirSync(dest, { recursive: true })
+    let entries = fs.readdirSync(src, { withFileTypes: true });
+
+    for (let entry of entries)
+    {
+        let srcPath = path.join(src, entry.name);
+        let destPath = path.join(dest, entry.name);
+
+        entry.isDirectory() ? copyDir(srcPath, destPath) : fs.copyFileSync(srcPath, destPath);
+    }
+}
+
+/**
+ * @remarks compresses a directory to a zip-like file
+ * @param dir the directory to compress
+ * @param zipPath the path the compressed directory should be written to
+ * @param callback the callback to run when the compression finishes
+ */
+export function archiveDirToZip(dir: string, zipPath: string, callback: Function) {
+    let output = fs.createWriteStream(zipPath);
+    let archive = archiver.default('zip', { zlib: { level: 9 } });
+   
+    output.on('close', () => {
+     callback();
+    });
+   
+    archive.pipe(output);
+    archive.directory(dir, '');
+    archive.finalize();
 }
